+++ Main/Features/CharacterCreation/ICareerMenuDataProvider.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using TAOM.Features.CharacterCreation.Models;
+
+namespace TAOM.Features.CharacterCreation;
+
+public interface ICareerMenuDataProvider
+{
+    IReadOnlyList<CareerMenuOptionDefinition> LoadCareerMenuOptions();
+    CareerMenuOptionDefinition GetOptionForCareer(string careerStringId);
+}