+++ Main/Features/CharacterCreation/ICharacterCreationContentService.cs
@@ -6,5 +6,6 @@ public interface ICharacterCreationContentService
 {
     void RegisterCustomCultures(CharacterCreationManager manager);
     void RegisterNarrativeMenus(CharacterCreationManager manager);
+    void RegisterCareerMenu(CharacterCreationManager manager);
     void OnCharacterCreationFinalize(CharacterCreationManager manager);
 }